Add test cases for the DB migration and initialization use cases
Two features are also introduced which facilitates testing.
First, configs now contain a role name suffix so distinct role names
may be produced to be used for different databases which are managed
concurrently by one shared database server.
Second, database authentication method is put in configs and a scram
hasher implementation is employed based on that setting, so role
passwords can be hashed before being sent to DBMS in an ALTER statement.